move exporter into an internal package, in preparation for profiles
